bookdown_site {bookdown}	R Documentation

R Markdown site generator for bookdown

Description

Implementation of custom R Markdown site generator for bookdown.

Usage

bookdown_site(input, ...)

Arguments

input
Website directory (or the name of a file within the directory).

...
Currently unused.

[Package bookdown version 0.17 Index]