build_chapter {bookdown}	R Documentation

Combine different parts of an HTML page

Description

Given the HTML header, body, and footer, etc, build an HTML page.

Usage

build_chapter(
  head,
  toc,
  chapter,
  link_prev,
  link_next,
  rmd_cur,
  html_cur,
  foot
)

Arguments

head
A character vector of the HTML code before the document title.

toc
A character vector of the table of contents.

chapter
The body of a chapter.

link_prev, link_next
The URL of the previous/next chapter (may be NULL).

rmd_cur
The Rmd filename of the current chapter (may be NULL).

html_cur
The HTML filename of the current chapter (may be NULL).

foot
A character vector of the HTML code after the chapter body.

Details

This function is for expert use only. The head and foot arguments may not be strictly the HTML header and footer. It depends on the HTML comment tokens in the template (see html_chapters).

[Package bookdown version 0.17 Index]