calibre {bookdown}	R Documentation

A wrapper function to convert e-books using Calibre

Description

This function calls the command ebook-convert in Calibre (http://calibre-ebook.com) to convert e-books.

Usage

calibre(input, output, options = "")

Arguments

input
The input filename.

output
The output filename or extension (if only an extension is provided, the output filename will be the input filename with its extension replaced by output; for example, calibre('foo.epub', 'mobi') generates ‘foo.mobi’).

options
A character vector of additional options to be passed to ebook-convert.

Value

The output filename.

[Package bookdown version 0.17 Index]