create_book {bookdown}	R Documentation

Create a bookdown project

Description

Create a bookdown project with multiple book output formats, including HTML. Choose one of two HTML book output formats:

Usage

create_gitbook(path)

create_bs4_book(path)

Arguments

path
Absolute path to an empty directory in which to create the bookdown project. In the RStudio IDE, if rstudioapi package available, an RStudio project will be created.

Details

Use create_gitbook() to use bookdown::gitbook() output format

Use create_bs4_book() to use a bookdown::bs4_book() output format

The function will create a folder with file structure for a bookdown project, and example files with information on how to start.

[Package bookdown version 0.36 Index]