gitbook {bookdown}	R Documentation

The GitBook output format

Description

This output format function ported a style provided by GitBook (https://www.gitbook.com) for R Markdown.

Usage

gitbook(
  fig_caption = TRUE,
  number_sections = TRUE,
  self_contained = FALSE,
  lib_dir = "libs",
  pandoc_args = NULL,
  ...,
  template = "default",
  split_by = c("chapter", "chapter+number", "section", "section+number", "rmd", "none"),
  split_bib = TRUE,
  config = list(),
  table_css = TRUE
)

Arguments

fig_caption, number_sections, self_contained, lib_dir, pandoc_args
... Arguments to be passed to rmarkdown::html_document() (... not including toc, and theme).

...
Other arguments to be passed to base_format. For html_book() and tufte_html_book(), ... is passed to html_chapters().

template
Pandoc template to use for rendering. Pass "default" to use the bookdown default template; pass a path to use a custom template. The default template should be sufficient for most use cases. In case you want to develop a custom template, we highly recommend to start from the default template: https://github.com/rstudio/bookdown/blob/master/inst/templates/gitbook.html.

split_by
How to name the HTML output files from the book: rmd uses the base filenames of the input Rmd files to create the HTML filenames, e.g. generate ‘chapter1.html’ for ‘chapter1.Rmd’; none means do not split the HTML file (the book will be a single HTML file); chapter means split the file by the first-level headers; section means the second-level headers. For chapter and section, the HTML filenames will be determined by the header ID's, e.g. the filename for the first chapter with a chapter title # Introduction will be ‘introduction.html’; for chapter+number and section+number, the chapter/section numbers will be prepended to the HTML filenames, e.g. ‘1-introduction.html’ and ‘2-1-literature.html’.

split_bib
Whether to split the bibliography onto separate pages where the citations are actually used.

config
A list of configuration options for the gitbook style, such as the font/theme settings.

table_css
TRUE to load gitbook's default CSS for tables. Choose FALSE to unload and use customized CSS (for exmaple, bootstrap) via the css option. Default is TRUE.

[Package bookdown version 0.17 Index]