html_document2 {bookdown}	R Documentation

Output formats that allow numbering and cross-referencing figures/tables/equations

Description

These are simple wrappers of the output format functions like rmarkdown::html_document(), and they added the capability of numbering figures/tables/equations/theorems and cross-referencing them. See References for the syntax. Note you can also cross-reference sections by their ID's using the same syntax when sections are numbered. In case you want to enable cross reference in other formats, use markdown_document2 with base_format argument.

Usage

html_document2(
  ...,
  number_sections = TRUE,
  pandoc_args = NULL,
  base_format = rmarkdown::html_document
)

html_fragment2(..., number_sections = FALSE)

html_notebook2(..., number_sections = FALSE)

html_vignette2(..., number_sections = FALSE)

ioslides_presentation2(..., number_sections = FALSE)

slidy_presentation2(..., number_sections = FALSE)

tufte_html2(..., number_sections = FALSE)

pdf_document2(...)

beamer_presentation2(..., number_sections = FALSE)

tufte_handout2(...)

tufte_book2(...)

markdown_document2(
  fig_caption = TRUE,
  md_extensions = NULL,
  pandoc_args = NULL,
  ...,
  base_format = rmarkdown::md_document
)

github_document2(...)

odt_document2(...)

powerpoint_presentation2(...)

rtf_document2(...)

word_document2(...)

Arguments

..., fig_caption, md_extensions, pandoc_args
Arguments to be passed to a specific output format function. For a function foo2(), its arguments are passed to foo(), e.g. ... of html_document2() are passed to rmarkdown::html_document().

number_sections
Whether to number section headers: if TRUE, figure/table numbers will be of the form X.i, where X is the current first-level section number, and i is an incremental number (the i-th figure/table); if FALSE, figures/tables will be numbered sequentially in the document from 1, 2, ..., and you cannot cross-reference section headers in this case.

base_format
An output format function to be used as the base format.

Value

An R Markdown output format object to be passed to rmarkdown::render().

Note

These function are expected to work with a single R Markdown document instead of multiple documents of a book, so they are to be passed to rmarkdown::render() instead of bookdown::render_book(). The functions tufte_*() are wrappers of funtions in the tufte package.

References

https://bookdown.org/yihui/bookdown/

[Package bookdown version 0.17 Index]