kindlegen {bookdown}	R Documentation

A wrapper function to convert EPUB to the Mobipocket format

Description

This function simply calls the command line tool kindlegen provided by Amazon to convert EPUB e-books to the Mobipocket format (‘.mobi’).

Usage

kindlegen(epub, exec = Sys.which("kindlegen"))

Arguments

epub
The path to a .epub file (e.g. created from the epub_book() format). If missing, it is automatically guessed from the book configurations.

exec
The path to the executable kindlegen, which can be downloaded from http://www.amazon.com/gp/feature.html?ie=UTF8&docId=1000765211.

Value

The path of the ‘.mobi’ file if the conversion is successful.

[Package bookdown version 0.17 Index]