publish_book {bookdown}	R Documentation

Publish a book to the web

Description

Publish a book to the web. Note that you should be sure to render all versions of the book before publishing, unless you have specified render = TRUE.

Usage

publish_book(
  name = NULL,
  account = NULL,
  server = NULL,
  render = c("none", "local", "server")
)

Arguments

name
Name of the book (this will be used in the URL path of the published book). Defaults to the book_filename in _bookdown.yml if not specified.

account
Account name to publish to. Will default to any previously published to account or any single account already associated with server.

server
Server to publish to (by default beta.rstudioconnect.com but any RStudio Connect server can be published to).

render
Rendering behavior for site: "none" to upload a static version of the current contents of the site directory; "local" to render the site locally then upload it; "server" to render the site on the server. Note that for "none" and "local" R scripts (.R) and markdown documents (.Rmd and .md) will not be uploaded to the server.

[Package bookdown version 0.17 Index]